sandbox: pull out shim manager from task manager
