+++ runtime/v2/manager.go
@@ -21,29 +21,19 @@ import (
 	"errors"
 	"fmt"
 	"os"
-	"os/exec"
 	"path/filepath"
-	"strings"
-	"sync"
 
 	"github.com/containerd/log"
 	"github.com/containerd/plugin"
 	"github.com/containerd/plugin/registry"
 
 	containerd "github.com/containerd/containerd/v2/client"
-	"github.com/containerd/containerd/v2/containers"
 	"github.com/containerd/containerd/v2/errdefs"
-	"github.com/containerd/containerd/v2/events/exchange"
 	"github.com/containerd/containerd/v2/metadata"
-	"github.com/containerd/containerd/v2/mount"
 	"github.com/containerd/containerd/v2/namespaces"
-	"github.com/containerd/containerd/v2/pkg/cleanup"
-	"github.com/containerd/containerd/v2/pkg/timeout"
 	"github.com/containerd/containerd/v2/platforms"
 	"github.com/containerd/containerd/v2/plugins"
-	"github.com/containerd/containerd/v2/protobuf"
 	"github.com/containerd/containerd/v2/runtime"
-	shimbinary "github.com/containerd/containerd/v2/runtime/v2/shim"
 )
 
 // Config for the v2 runtime
@@ -55,37 +45,6 @@ type Config struct {
 }
 
 func init() {
-	// Task manager uses shim manager as a dependency to manage shim instances.
-	// However, due to time limits and to avoid migration steps in 1.6 release,
-	// use the following workaround.
-	// This expected to be removed in 1.7.
-	registry.Register(&plugin.Registration{
-		Type: plugins.ShimPlugin,
-		ID:   "shim",
-		Requires: []plugin.Type{
-			plugins.EventPlugin,
-			plugins.MetadataPlugin,
-		},
-		InitFn: func(ic *plugin.InitContext) (interface{}, error) {
-			m, err := ic.GetSingle(plugins.MetadataPlugin)
-			if err != nil {
-				return nil, err
-			}
-			ep, err := ic.GetByID(plugins.EventPlugin, "exchange")
-			if err != nil {
-				return nil, err
-			}
-			events := ep.(*exchange.Exchange)
-			cs := metadata.NewContainerStore(m.(*metadata.DB))
-			return NewShimManager(&ManagerConfig{
-				Address:      ic.Properties[plugins.PropertyGRPCAddress],
-				TTRPCAddress: ic.Properties[plugins.PropertyTTRPCAddress],
-				Events:       events,
-				Store:        cs,
-			})
-		},
-	})
-
 	registry.Register(&plugin.Registration{
 		Type: plugins.RuntimePluginV2,
 		ID:   "task",
@@ -143,320 +102,6 @@ func init() {
 	})
 }
 
-type ManagerConfig struct {
-	Store        containers.Store
-	Events       *exchange.Exchange
-	Address      string
-	TTRPCAddress string
-	SchedCore    bool
-}
-
-// NewShimManager creates a manager for v2 shims
-func NewShimManager(config *ManagerConfig) (*ShimManager, error) {
-	m := &ShimManager{
-		containerdAddress:      config.Address,
-		containerdTTRPCAddress: config.TTRPCAddress,
-		shims:                  runtime.NewNSMap[ShimInstance](),
-		events:                 config.Events,
-		containers:             config.Store,
-		schedCore:              config.SchedCore,
-	}
-
-	return m, nil
-}
-
-// ShimManager manages currently running shim processes.
-// It is mainly responsible for launching new shims and for proper shutdown and cleanup of existing instances.
-// The manager is unaware of the underlying services shim provides and lets higher level services consume them,
-// but don't care about lifecycle management.
-type ShimManager struct {
-	containerdAddress      string
-	containerdTTRPCAddress string
-	schedCore              bool
-	shims                  *runtime.NSMap[ShimInstance]
-	events                 *exchange.Exchange
-	containers             containers.Store
-	// runtimePaths is a cache of `runtime names` -> `resolved fs path`
-	runtimePaths sync.Map
-}
-
-// ID of the shim manager
-func (m *ShimManager) ID() string {
-	return plugins.RuntimePluginV2.String() + ".shim"
-}
-
-// Start launches a new shim instance
-func (m *ShimManager) Start(ctx context.Context, id string, bundle *Bundle, opts runtime.CreateOpts) (_ ShimInstance, retErr error) {
-	shim, err := m.startShim(ctx, bundle, id, opts)
-	if err != nil {
-		return nil, err
-	}
-	defer func() {
-		if retErr != nil {
-			m.cleanupShim(ctx, shim)
-		}
-	}()
-
-	if err := m.shims.Add(ctx, shim); err != nil {
-		return nil, fmt.Errorf("failed to add task: %w", err)
-	}
-
-	return shim, nil
-}
-
-func (m *ShimManager) startShim(ctx context.Context, bundle *Bundle, id string, opts runtime.CreateOpts) (*shim, error) {
-	ns, err := namespaces.NamespaceRequired(ctx)
-	if err != nil {
-		return nil, err
-	}
-	ctx = log.WithLogger(ctx, log.G(ctx).WithField("namespace", ns))
-
-	topts := opts.TaskOptions
-	if topts == nil || topts.GetValue() == nil {
-		topts = opts.RuntimeOptions
-	}
-
-	runtimePath, err := m.resolveRuntimePath(opts.Runtime)
-	if err != nil {
-		return nil, fmt.Errorf("failed to resolve runtime path: %w", err)
-	}
-
-	b := shimBinary(bundle, shimBinaryConfig{
-		runtime:      runtimePath,
-		address:      m.containerdAddress,
-		ttrpcAddress: m.containerdTTRPCAddress,
-		schedCore:    m.schedCore,
-	})
-	shim, err := b.Start(ctx, protobuf.FromAny(topts), func() {
-		log.G(ctx).WithField("id", id).Info("shim disconnected")
-
-		cleanupAfterDeadShim(cleanup.Background(ctx), id, m.shims, m.events, b)
-		// Remove self from the runtime task list. Even though the cleanupAfterDeadShim()
-		// would publish taskExit event, but the shim.Delete() would always failed with ttrpc
-		// disconnect and there is no chance to remove this dead task from runtime task lists.
-		// Thus it's better to delete it here.
-		m.shims.Delete(ctx, id)
-	})
-	if err != nil {
-		return nil, fmt.Errorf("start failed: %w", err)
-	}
-
-	return shim, nil
-}
-
-// restoreBootstrapParams reads bootstrap.json to restore shim configuration.
-// If its an old shim, this will perform migration - read address file and write default bootstrap
-// configuration (version = 2, protocol = ttrpc, and address).
-func restoreBootstrapParams(bundlePath string) (shimbinary.BootstrapParams, error) {
-	filePath := filepath.Join(bundlePath, "bootstrap.json")
-
-	// Read bootstrap.json if exists
-	if _, err := os.Stat(filePath); err == nil {
-		return readBootstrapParams(filePath)
-	} else if !errors.Is(err, os.ErrNotExist) {
-		return shimbinary.BootstrapParams{}, fmt.Errorf("failed to stat %s: %w", filePath, err)
-	}
-
-	// File not found, likely its an older shim. Try migrate.
-
-	address, err := shimbinary.ReadAddress(filepath.Join(bundlePath, "address"))
-	if err != nil {
-		return shimbinary.BootstrapParams{}, fmt.Errorf("unable to migrate shim: failed to get socket address for bundle %s: %w", bundlePath, err)
-	}
-
-	params := shimbinary.BootstrapParams{
-		Version:  2,
-		Address:  address,
-		Protocol: "ttrpc",
-	}
-
-	if err := writeBootstrapParams(filePath, params); err != nil {
-		return shimbinary.BootstrapParams{}, fmt.Errorf("unable to migrate: failed to write bootstrap.json file: %w", err)
-	}
-
-	return params, nil
-}
-
-func (m *ShimManager) resolveRuntimePath(runtime string) (string, error) {
-	if runtime == "" {
-		return "", fmt.Errorf("no runtime name")
-	}
-
-	// Custom path to runtime binary
-	if filepath.IsAbs(runtime) {
-		// Make sure it exists before returning ok
-		if _, err := os.Stat(runtime); err != nil {
-			return "", fmt.Errorf("invalid custom binary path: %w", err)
-		}
-
-		return runtime, nil
-	}
-
-	// Check if relative path to runtime binary provided
-	if strings.Contains(runtime, "/") {
-		return "", fmt.Errorf("invalid runtime name %s, correct runtime name should be either format like `io.containerd.runc.v2` or a full path to the binary", runtime)
-	}
-
-	// Preserve existing logic and resolve runtime path from runtime name.
-
-	name := shimbinary.BinaryName(runtime)
-	if name == "" {
-		return "", fmt.Errorf("invalid runtime name %s, correct runtime name should be either format like `io.containerd.runc.v2` or a full path to the binary", runtime)
-	}
-
-	if path, ok := m.runtimePaths.Load(name); ok {
-		return path.(string), nil
-	}
-
-	var (
-		cmdPath string
-		lerr    error
-	)
-
-	binaryPath := shimbinary.BinaryPath(runtime)
-	if _, serr := os.Stat(binaryPath); serr == nil {
-		cmdPath = binaryPath
-	}
-
-	if cmdPath == "" {
-		if cmdPath, lerr = exec.LookPath(name); lerr != nil {
-			if eerr, ok := lerr.(*exec.Error); ok {
-				if eerr.Err == exec.ErrNotFound {
-					self, err := os.Executable()
-					if err != nil {
-						return "", err
-					}
-
-					// Match the calling binaries (containerd) path and see
-					// if they are side by side. If so, execute the shim
-					// found there.
-					testPath := filepath.Join(filepath.Dir(self), name)
-					if _, serr := os.Stat(testPath); serr == nil {
-						cmdPath = testPath
-					}
-					if cmdPath == "" {
-						return "", fmt.Errorf("runtime %q binary not installed %q: %w", runtime, name, os.ErrNotExist)
-					}
-				}
-			}
-		}
-	}
-
-	cmdPath, err := filepath.Abs(cmdPath)
-	if err != nil {
-		return "", err
-	}
-
-	if path, ok := m.runtimePaths.LoadOrStore(name, cmdPath); ok {
-		// We didn't store cmdPath we loaded an already cached value. Use it.
-		cmdPath = path.(string)
-	}
-
-	return cmdPath, nil
-}
-
-// cleanupShim attempts to properly delete and cleanup shim after error
-func (m *ShimManager) cleanupShim(ctx context.Context, shim *shim) {
-	dctx, cancel := timeout.WithContext(cleanup.Background(ctx), cleanupTimeout)
-	defer cancel()
-
-	_ = shim.Delete(dctx)
-	m.shims.Delete(dctx, shim.ID())
-}
-
-func (m *ShimManager) Get(ctx context.Context, id string) (ShimInstance, error) {
-	return m.shims.Get(ctx, id)
-}
-
-// Delete a runtime task
-func (m *ShimManager) Delete(ctx context.Context, id string) error {
-	shim, err := m.shims.Get(ctx, id)
-	if err != nil {
-		return err
-	}
-
-	err = shim.Delete(ctx)
-	m.shims.Delete(ctx, id)
-
-	return err
-}
-
-func (m *ShimManager) Load(ctx context.Context, bundle *Bundle) error {
-	var (
-		runtime string
-		id      = bundle.ID
-	)
-
-	// If we're on 1.6+ and specified custom path to the runtime binary, path will be saved in 'shim-binary-path' file.
-	if data, err := os.ReadFile(filepath.Join(bundle.Path, "shim-binary-path")); err == nil {
-		runtime = string(data)
-	} else if err != nil && !os.IsNotExist(err) {
-		log.G(ctx).WithError(err).Error("failed to read `runtime` path from bundle")
-	}
-
-	// Query runtime name from metadata store
-	if runtime == "" {
-		container, err := m.containers.Get(ctx, id)
-		if err != nil {
-			log.G(ctx).WithError(err).Errorf("loading container %s", id)
-			if umountErr := mount.UnmountRecursive(filepath.Join(bundle.Path, "rootfs"), 0); umountErr != nil {
-				log.G(ctx).WithError(umountErr).Errorf("failed to unmount of rootfs %s", id)
-			}
-			return err
-		}
-		runtime = container.Runtime.Name
-	}
-
-	runtime, err := m.resolveRuntimePath(runtime)
-	if err != nil {
-		return fmt.Errorf("failed to resolve runtime path: %w", err)
-	}
-
-	binaryCall := shimBinary(bundle,
-		shimBinaryConfig{
-			runtime:      runtime,
-			address:      m.containerdAddress,
-			ttrpcAddress: m.containerdTTRPCAddress,
-			schedCore:    m.schedCore,
-		})
-	shim, err := loadShimTask(ctx, bundle, func() {
-		log.G(ctx).WithField("id", id).Info("shim disconnected")
-
-		cleanupAfterDeadShim(cleanup.Background(ctx), id, m.shims, m.events, binaryCall)
-		// Remove self from the runtime task list.
-		m.shims.Delete(ctx, id)
-	})
-	if err != nil {
-		cleanupAfterDeadShim(ctx, id, m.shims, m.events, binaryCall)
-		return fmt.Errorf("unable to load shim %q: %w", id, err)
-	}
-
-	// There are 2 possibilities for the loaded shim here:
-	// 1. It could be a shim that is running a task.
-	// 3. Or it could be a shim that was created for running a task but
-	// something happened (probably a containerd crash) and the task was never
-	// created. This shim process should be cleaned up here. Look at
-	// containerd/containerd#6860 for further details.
-	pInfo, pidErr := shim.Pids(ctx)
-	if len(pInfo) == 0 || errors.Is(pidErr, errdefs.ErrNotFound) {
-		log.G(ctx).WithField("id", id).Info("cleaning leaked shim process")
-		// We are unable to get Pids from the shim, we should clean it up here.
-		// No need to do anything for removeTask since we never added this shim.
-		shim.delete(ctx, func(ctx context.Context, id string) {})
-	} else {
-		m.shims.Add(ctx, shim.ShimInstance)
-	}
-	return nil
-}
-
-func (m *ShimManager) LoadShim(ctx context.Context, bundle *Bundle, onClose func()) error {
-	shim, err := loadShim(ctx, bundle, onClose)
-	if err != nil {
-		return err
-	}
-	return m.shims.Add(ctx, shim)
-}
-
 // TaskManager wraps task service client on top of shim manager.
 type TaskManager struct {
 	root                 string
@@ -503,42 +148,7 @@ func (m *TaskManager) Create(ctx context.Context, taskID string, opts runtime.Cr
 			return t, err
 		}
 	}
-
-	shim, err := m.shimManager.Start(ctx, taskID, bundle, opts)
-	if err != nil {
-		return nil, fmt.Errorf("failed to start shim: %w", err)
-	}
-
-	// Cast to shim task and call task service to create a new container task instance.
-	// This will not be required once shim service / client implemented.
-	shimTask, err := newShimTask(shim)
-	if err != nil {
-		return nil, err
-	}
-
-	task, err := shimTask.Create(ctx, opts)
-	if err != nil {
-		// NOTE: ctx contains required namespace information.
-		m.shimManager.shims.Delete(ctx, taskID)
-
-		dctx, cancel := timeout.WithContext(cleanup.Background(ctx), cleanupTimeout)
-		defer cancel()
-
-		_, errShim := shimTask.delete(dctx, func(context.Context, string) {})
-		if errShim != nil {
-			if errdefs.IsDeadlineExceeded(errShim) {
-				dctx, cancel = timeout.WithContext(cleanup.Background(ctx), cleanupTimeout)
-				defer cancel()
-			}
-
-			shimTask.Shutdown(dctx)
-			shimTask.Close()
-		}
-
-		return nil, fmt.Errorf("failed to create shim task: %w", err)
-	}
-
-	return task, nil
+	return m.shimManager.Create(ctx, taskID, bundle, opts)
 }
 
 // Get a specific task
